Move api doc export to own module
